refactor(api): merge robot related types
