--- pkg/api/roboscale.io/v1alpha1/robotartifact_types.go
@@ -1,45 +1,0 @@
-/*
-Copyright 2022.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package v1alpha1
-
-import (
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-//+kubebuilder:object:root=true
-//+kubebuilder:subresource:status
-
-// RobotArtifact is the Schema for the robotartifacts API
-type RobotArtifact struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-
-	Template RobotSpec `json:"template,omitempty"`
-}
-
-//+kubebuilder:object:root=true
-
-// RobotArtifactList contains a list of RobotArtifact
-type RobotArtifactList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []RobotArtifact `json:"items"`
-}
-
-func init() {
-	SchemeBuilder.Register(&RobotArtifact{}, &RobotArtifactList{})
-}
